Rename component folder to components
